+++ docs/examples/opentelemetry-example-app/src/opentelemetry_example_app/grpc/hello_world_client.py
@@ -17,11 +17,11 @@
 
 """The Python implementation of the GRPC helloworld.Greeter client.
 
-Note that you need ``opentelemetry-instrumentation-grpc`` and ``protobuf`` to be installed
+Note that you need ``opentelemetry-ext-grpc`` and ``protobuf`` to be installed
 to run these examples. To run this script in the context of the example app,
 install ``opentelemetry-example-app``::
 
-    pip install -e ext/opentelemetry-instrumentation-grpc/
+    pip install -e ext/opentelemetry-ext-grpc/
     pip install -e docs/examples/opentelemetry-example-app
 
 Then run the server in one shell::
@@ -42,8 +42,8 @@ import logging
 import grpc
 
 from opentelemetry import trace
-from opentelemetry.instrumentation.grpc import client_interceptor
-from opentelemetry.instrumentation.grpc.grpcext import intercept_channel
+from opentelemetry.ext.grpc import client_interceptor
+from opentelemetry.ext.grpc.grpcext import intercept_channel
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import (
     ConsoleSpanExporter,