+++ ext/opentelemetry-ext-grpc/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Change package name to opentelemetry-instrumentation-grpc
+- Change package name to opentelemetry-ext-grpc
   ([#961](https://github.com/open-telemetry/opentelemetry-python/pull/961))
 
 ## Version 0.11b0